TPS-3: Create basic entity and repo classes

+++ loan-service/src/main/java/repository/LoanRepository.java
@@ -6,14 +6,11 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-import java.time.LocalDateTime;
-import java.util.Collections;
 import java.util.List;
 
 @Repository
-public interface LoanRepository extends JpaRepository<Loan, BigInteger> {
+public interface LoanRepository extends JpaRepository<Loan, Long> {
 
-    @Query("SELECT l FROM Loan l WHERE l.customerId = :customerId" )
+    @Query("SELECT l FROM Loan l WHERE l.customerId = :customerId")
     List<Loan> getLoansByCustomerId(@Param("customerId") Long customerId);
 }
